Rename download link components for consistency
